Refactor `package.json` (#322)
This PR does a few things:

Run the sort-package-json tool to put the package.json keys into their canonical ordering
Move nyc configuration into a standalone config file, like Prettier, ESLint, and all the other tools.
Also, I changed the color highlighting levels of NYC. Now, it will be:

Red for < 80% coverage
Yellow for 80% - 95% coverage
Green for 95+% coverage
This is what I used for color highlighting in PayID, but I'm happy to change the cutoffs to align with CodeCov configuration. It looked like the target was 80%, so that seemed like a reasonable lower bound.